+++ src/views/branches/branchListingView.ts
@@ -1,16 +1,11 @@
-import { View } from '../general/view';
 import { Ref } from '../../typings/git';
 import { TextView } from '../general/textView';
 
-export class BranchListingView extends View {
+export class BranchListingView extends TextView {
 
-  get id() { return this.branch.name?.toString() + this.branch.type.toString(); }
+  get id() { return this.ref.name?.toString() + this.ref.type.toString(); }
 
-  constructor(public branch: Ref, active = false) {
-    super();
-
-    this.addSubview(
-      new TextView(`${active ? '*' : ' '} ${branch.name}`)
-    );
+  constructor(public ref: Ref, active = false) {
+    super(`${active ? '*' : ' '} ${ref.name}`);
   }
 }