+++ src/views/tags/tagListingView.ts
@@ -1,16 +1,11 @@
-import { View } from '../general/view';
 import { Ref } from '../../typings/git';
 import { TextView } from '../general/textView';
 
-export class TagListingView extends View {
+export class TagListingView extends TextView {
 
-  get id() { return this.tag.name?.toString() + this.tag.type.toString(); }
+  get id() { return this.ref.name?.toString() + this.ref.type.toString(); }
 
-  constructor(public tag: Ref) {
-    super();
-
-    this.addSubview(
-      new TextView(`  ${tag.name}`)
-    );
+  constructor(public ref: Ref) {
+    super(`  ${ref.name}`);
   }
 }